+++ efpmd/src/hess.c
@@ -31,14 +31,14 @@
 
 void sim_hess(struct efp *, const struct cfg *, const struct sys *);
 
-static void compute_gradient(struct efp *efp, int n_frags,
+static void compute_gradient(struct efp *efp, size_t n_frags,
 		const double *xyzabc, double *grad)
 {
 	check_fail(efp_set_coordinates(efp, EFP_COORD_TYPE_XYZABC, xyzabc));
 	check_fail(efp_compute(efp, 1));
 	check_fail(efp_get_gradient(efp, n_frags, grad));
 
-	for (int i = 0; i < n_frags; i++) {
+	for (size_t i = 0; i < n_frags; i++) {
 		const double *euler = xyzabc + 6 * i + 3;
 		double *gradptr = grad + 6 * i + 3;
 
@@ -46,15 +46,15 @@ static void compute_gradient(struct efp *efp, int n_frags,
 	}
 }
 
-static void show_progress(int disp, int total, const char *dir)
+static void show_progress(size_t disp, size_t total, const char *dir)
 {
-	printf("COMPUTING DISPLACEMENT %4d OF %d (%s)\n", disp, total, dir);
+	printf("COMPUTING DISPLACEMENT %4zu OF %zu (%s)\n", disp, total, dir);
 	fflush(stdout);
 }
 
 static void compute_hessian(struct efp *efp, const struct cfg *cfg, double *hess)
 {
-	int n_frags, n_coord;
+	size_t n_frags, n_coord;
 	double *xyzabc, *grad_f, *grad_b;
 	bool central = cfg_get_bool(cfg, "hess_central");
 
@@ -70,7 +70,7 @@ static void compute_hessian(struct efp *efp, const struct cfg *cfg, double *hess
 	if (!central) {
 		check_fail(efp_get_gradient(efp, n_frags, grad_b));
 
-		for (int i = 0; i < n_frags; i++) {
+		for (size_t i = 0; i < n_frags; i++) {
 			const double *euler = xyzabc + 6 * i + 3;
 			double *gradptr = grad_b + 6 * i + 3;
 
@@ -78,7 +78,7 @@ static void compute_hessian(struct efp *efp, const struct cfg *cfg, double *hess
 		}
 	}
 
-	for (int i = 0; i < n_coord; i++) {
+	for (size_t i = 0; i < n_coord; i++) {
 		double save = xyzabc[i];
 		double step = i % 6 < 3 ? cfg_get_double(cfg, "hess_step_dist") :
 				cfg_get_double(cfg, "hess_step_angle");
@@ -95,7 +95,7 @@ static void compute_hessian(struct efp *efp, const struct cfg *cfg, double *hess
 
 		double delta = central ? 2.0 * step : step;
 
-		for (int j = 0; j < n_coord; j++)
+		for (size_t j = 0; j < n_coord; j++)
 			hess[i * n_coord + j] = (grad_f[j] - grad_b[j]) / delta;
 
 		xyzabc[i] = save;
@@ -105,8 +105,8 @@ static void compute_hessian(struct efp *efp, const struct cfg *cfg, double *hess
 	check_fail(efp_set_coordinates(efp, EFP_COORD_TYPE_XYZABC, xyzabc));
 
 	/* reduce error by computing the average of H(i,j) and H(j,i) */
-	for (int i = 0; i < n_coord; i++) {
-		for (int j = i + 1; j < n_coord; j++) {
+	for (size_t i = 0; i < n_coord; i++) {
+		for (size_t j = i + 1; j < n_coord; j++) {
 			double sum = hess[i * n_coord + j] + hess[j * n_coord + i];
 
 			hess[i * n_coord + j] = 0.5 * sum;
@@ -126,14 +126,14 @@ static void get_inertia_factor(const double *inertia, const mat_t *rotmat,
 {
 	double fact[3];
 
-	for (int i = 0; i < 3; i++)
+	for (size_t i = 0; i < 3; i++)
 		fact[i] = inertia[i] < EPSILON ? 0.0 : 1.0 / sqrt(inertia[i]);
 
-	for (int i = 0; i < 3; i++) {
-		for (int j = 0; j < 3; j++) {
+	for (size_t i = 0; i < 3; i++) {
+		for (size_t j = 0; j < 3; j++) {
 			double sum = 0.0;
 
-			for (int k = 0; k < 3; k++)
+			for (size_t k = 0; k < 3; k++)
 				sum += fact[k] * mat_get(rotmat, i, k) * mat_get(rotmat, j, k);
 
 			mat_set(inertia_fact, i, j, sum);
@@ -144,13 +144,13 @@ static void get_inertia_factor(const double *inertia, const mat_t *rotmat,
 static void get_weight_factor(struct efp *efp, double *mass_fact,
 				mat_t *inertia_fact)
 {
-	int n_frags;
+	size_t n_frags;
 	check_fail(efp_get_frag_count(efp, &n_frags));
 
 	double xyzabc[6 * n_frags];
 	check_fail(efp_get_coordinates(efp, n_frags, xyzabc));
 
-	for (int i = 0; i < n_frags; i++) {
+	for (size_t i = 0; i < n_frags; i++) {
 		double mass;
 		check_fail(efp_get_frag_mass(efp, i, &mass));
 
@@ -169,22 +169,22 @@ static void get_weight_factor(struct efp *efp, double *mass_fact,
 	}
 }
 
-static void w_tr_tr(double fact1, double fact2, int stride,
+static void w_tr_tr(double fact1, double fact2, size_t stride,
 			const double *in, double *out)
 {
-	for (int i = 0; i < 3; i++)
-		for (int j = 0; j < 3; j++)
+	for (size_t i = 0; i < 3; i++)
+		for (size_t j = 0; j < 3; j++)
 			out[stride * i + j] = fact1 * fact2 * in[stride * i + j];
 }
 
-static void w_tr_rot(double fact1, const mat_t *fact2, int stride,
+static void w_tr_rot(double fact1, const mat_t *fact2, size_t stride,
 			const double *in, double *out)
 {
-	for (int i = 0; i < 3; i++) {
-		for (int j = 0; j < 3; j++) {
+	for (size_t i = 0; i < 3; i++) {
+		for (size_t j = 0; j < 3; j++) {
 			double w = 0.0;
 
-			for (int k = 0; k < 3; k++)
+			for (size_t k = 0; k < 3; k++)
 				w += mat_get(fact2, j, k) * in[stride * i + k];
 
 			out[stride * i + j] = w * fact1;
@@ -192,14 +192,14 @@ static void w_tr_rot(double fact1, const mat_t *fact2, int stride,
 	}
 }
 
-static void w_rot_tr(const mat_t *fact1, double fact2, int stride,
+static void w_rot_tr(const mat_t *fact1, double fact2, size_t stride,
 			const double *in, double *out)
 {
-	for (int i = 0; i < 3; i++) {
-		for (int j = 0; j < 3; j++) {
+	for (size_t i = 0; i < 3; i++) {
+		for (size_t j = 0; j < 3; j++) {
 			double w = 0.0;
 
-			for (int k = 0; k < 3; k++)
+			for (size_t k = 0; k < 3; k++)
 				w += mat_get(fact1, i, k) * in[stride * k + j];
 
 			out[stride * i + j] = w * fact2;
@@ -207,17 +207,17 @@ static void w_rot_tr(const mat_t *fact1, double fact2, int stride,
 	}
 }
 
-static void w_rot_rot(const mat_t *fact1, const mat_t *fact2, int stride,
+static void w_rot_rot(const mat_t *fact1, const mat_t *fact2, size_t stride,
 			const double *in, double *out)
 {
-	for (int i = 0; i < 3; i++) {
-		for (int j = 0; j < 3; j++) {
+	for (size_t i = 0; i < 3; i++) {
+		for (size_t j = 0; j < 3; j++) {
 			double w1 = 0.0;
 
-			for (int ii = 0; ii < 3; ii++) {
+			for (size_t ii = 0; ii < 3; ii++) {
 				double w2 = 0.0;
 
-				for (int jj = 0; jj < 3; jj++)
+				for (size_t jj = 0; jj < 3; jj++)
 					w2 += mat_get(fact2, j, jj) * in[stride * ii + jj];
 
 				w1 += w2 * mat_get(fact1, i, ii);
@@ -230,7 +230,7 @@ static void w_rot_rot(const mat_t *fact1, const mat_t *fact2, int stride,
 
 static void mass_weight_hessian(struct efp *efp, const double *in, double *out)
 {
-	int n_frags, n_coord;
+	size_t n_frags, n_coord;
 
 	check_fail(efp_get_frag_count(efp, &n_frags));
 	n_coord = 6 * n_frags;
@@ -240,9 +240,9 @@ static void mass_weight_hessian(struct efp *efp, const double *in, double *out)
 
 	get_weight_factor(efp, mass_fact, inertia_fact);
 
-	for (int i = 0; i < n_frags; i++) {
-		for (int j = 0; j < n_frags; j++) {
-			int offset = 6 * n_coord * i + 6 * j;
+	for (size_t i = 0; i < n_frags; i++) {
+		for (size_t j = 0; j < n_frags; j++) {
+			size_t offset = 6 * n_coord * i + 6 * j;
 
 			w_tr_tr(mass_fact[i], mass_fact[j], n_coord,
 					in + offset,
@@ -263,7 +263,7 @@ static void mass_weight_hessian(struct efp *efp, const double *in, double *out)
 	}
 }
 
-static void print_mode(int mode, double eigen)
+static void print_mode(size_t mode, double eigen)
 {
 	/* preserve sign for imaginary frequencies */
 	eigen = copysign(sqrt(fabs(eigen)), eigen);
@@ -271,7 +271,7 @@ static void print_mode(int mode, double eigen)
 	/* convert to cm-1 */
 	eigen = FINE_CONST / 2.0 / PI / BOHR_RADIUS / sqrt(AMU_TO_AU) * 1.0e8 * eigen;
 
-	printf("    MODE %4d    FREQUENCY %10.3lf cm-1\n\n", mode, eigen);
+	printf("    MODE %4zu    FREQUENCY %10.3lf cm-1\n\n", mode, eigen);
 }
 
 void sim_hess(struct efp *efp, const struct cfg *cfg, const struct sys *sys)
@@ -285,7 +285,7 @@ void sim_hess(struct efp *efp, const struct cfg *cfg, const struct sys *sys)
 	print_energy(efp);
 	print_gradient(efp);
 
-	int n_frags, n_coord;
+	size_t n_frags, n_coord;
 	double *hess, *mass_hess, *eigen;
 
 	check_fail(efp_get_frag_count(efp, &n_frags));
@@ -307,10 +307,10 @@ void sim_hess(struct efp *efp, const struct cfg *cfg, const struct sys *sys)
 
 	eigen = xmalloc(n_coord * sizeof(double));
 
-	if (c_dsyev('V', 'U', n_coord, mass_hess, n_coord, eigen))
+	if (c_dsyev('V', 'U', (int)n_coord, mass_hess, (int)n_coord, eigen))
 		error("UNABLE TO DIAGONALIZE MASS-WEIGHTED HESSIAN MATRIX");
 
-	for (int i = 0; i < n_coord; i++) {
+	for (size_t i = 0; i < n_coord; i++) {
 		print_mode(i + 1, eigen[i]);
 		print_vector(n_coord, mass_hess + i * n_coord);
 	}
