+++ tests/qm_1.c
@@ -31,7 +31,7 @@ static const double qm_xyz[] = {
 	BOHR( 4.9), BOHR( 0.0), BOHR( 4.7)
 };
 
-static enum efp_result get_electron_density_field(int n_pt,
+static enum efp_result get_electron_density_field(size_t n_pt,
 		const double *xyz, double *field, void *user_data)
 {
 	(void)xyz;
